backend/requirements.txt
fastapi[standard]
uvicorn[standard]
python-dotenv
SQLAlchemy
asyncpg
sqlalchemy[asyncio]
psycopg2-binary
python-jose[cryptography]
passlib[bcrypt]
pydantic
psycopg2-binary
pyjwt
